Switch to model-based Illusts
I noticed we spent quite a bit of time re-initializing all of the
illustration delegates and dealing with bindings, since all of the
fetching is handled on the QML side. That ends up sending things over
the engine barrier (slow), creates bindings (also slow) and re-creates
all of the existing illustration delegates (even slower.)

We can avoid a lot of these problems by using Qt models, which is added
in Piqi. It's handled C++ side so it's much faster, although the
fetching API is called in QML for now.